+++ src/main/java/com/kustacks/kuring/notice/presentation/NoticeControllerV1.java
@@ -1,9 +1,9 @@
 package com.kustacks.kuring.notice.presentation;
 
 import com.kustacks.kuring.category.business.CategoryService;
-import com.kustacks.kuring.category.common.dto.CategoryListResponse;
-import com.kustacks.kuring.category.common.dto.SubscribeCategoriesV1Request;
-import com.kustacks.kuring.category.common.dto.SubscribeCategoriesResponse;
+import com.kustacks.kuring.notice.presentation.dto.CategoryListResponse;
+import com.kustacks.kuring.notice.presentation.dto.SubscribeCategoriesV1Request;
+import com.kustacks.kuring.notice.presentation.dto.SubscribeCategoriesResponse;
 import com.kustacks.kuring.notice.domain.CategoryName;
 import com.kustacks.kuring.common.firebase.FirebaseService;
 import com.kustacks.kuring.notice.business.NoticeService;
