+++ src/main/java/com/kustacks/kuring/notice/presentation/dto/CategoryListResponse.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.category.common.dto;
+package com.kustacks.kuring.notice.presentation.dto;
 
 import com.kustacks.kuring.common.dto.ResponseDto;
 import lombok.AccessLevel;
